+++ src/main/java/com/answerdigital/benhession/academy/answerkingweek2/repositories/OrderRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface OrderRepository extends CrudRepository<Order, Integer> {
+public interface OrderRepository extends CrudRepository<Order, Long>
+{
     List<Order> findAll();
 }
